+++ src/main/java/roomescape/infrastructure/MemberId.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.infrastructure;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
